+++ packages/rainbowkit/src/components/Dialog/Dialog.css.ts
@@ -19,6 +19,7 @@ const nestedModalZIndexes: Record<string, number> = {
 const bleed = 200;
 export const overlay = style([
   sprinkles({
+    backdropFilter: 'modalOverlay',
     background: 'modalBackdrop',
     display: 'flex',
     justifyContent: 'center',