+++ packages/rainbowkit/src/css/sprinkles.css.ts
@@ -59,6 +59,9 @@ const themeContractValues = {
     selectedWallet: '',
     walletLogo: '',
   },
+  blurs: {
+    modalOverlay: '',
+  },
 };
 
 export type ThemeVars = typeof themeContractValues;
@@ -116,36 +119,6 @@ const dimensions = {
 const flexAlignment = ['flex-start', 'flex-end', 'center'] as const;
 
 const textAlignments = ['left', 'center', 'inherit'] as const;
-
-export const growTransforms = {
-  grow: 'scale(1.025)',
-  growLg: 'scale(1.1)',
-} as const;
-
-const shrinkTransforms = {
-  shrink: 'scale(0.95)',
-  shrinkSm: 'scale(0.9)',
-} as const;
-
-const transforms = {
-  ...growTransforms,
-  ...shrinkTransforms,
-} as const;
-
-const interactionProperties = defineProperties({
-  conditions: {
-    base: {},
-    hover: { selector: '&:hover' },
-    active: { selector: '&:active' },
-  },
-  defaultCondition: 'base',
-  properties: {
-    transform: transforms,
-    transition: {
-      default: '0.125s ease',
-    },
-  },
-});
 
 export const largeScreenMinWidth = 768;
 
@@ -230,8 +203,14 @@ const unresponsiveProperties = defineProperties({
     right: {
       '0': '0',
     },
+    transition: {
+      default: '0.125s ease',
+    },
     userSelect: ['none'] as const,
     width: dimensions,
+    backdropFilter: {
+      ...themeVars.blurs,
+    },
   } as const,
   shorthands: {
     margin: ['marginTop', 'marginBottom', 'marginLeft', 'marginRight'],
@@ -260,7 +239,6 @@ const colorProperties = defineProperties({
 
 export const sprinkles = createSprinkles(
   colorProperties,
-  interactionProperties,
   responsiveProperties,
   unresponsiveProperties
 );