+++ packages/rainbowkit/src/utils/isMobile.ts
@@ -1,17 +1,23 @@
 export function isAndroid(): boolean {
   return (
-    typeof navigator !== 'undefined' &&
-    /Android\s([0-9.]+)/.test(navigator.userAgent) // Source: https://github.com/DamonOehlman/detect-browser/blob/master/src/index.ts
+    typeof navigator !== 'undefined' && /android/i.test(navigator.userAgent)
   );
+}
+
+export function isSmallIOS(): boolean {
+  return (
+    typeof navigator !== 'undefined' && /iPhone|iPod/.test(navigator.userAgent)
+  );
+}
+
+export function isLargeIOS(): boolean {
+  return typeof navigator !== 'undefined' && /iPad/.test(navigator.userAgent);
 }
 
 export function isIOS(): boolean {
-  return (
-    typeof navigator !== 'undefined' &&
-    /Version\/([0-9._]+).*Mobile.*Safari.*/.test(navigator.userAgent) // Source: https://github.com/DamonOehlman/detect-browser/blob/master/src/index.ts
-  );
+  return isSmallIOS() || isLargeIOS();
 }
 
 export function isMobile(): boolean {
-  return isAndroid() || isIOS();
+  return isAndroid() || isSmallIOS();
 }