+++ packages/example/CHANGELOG.md
@@ -1,5 +1,23 @@
 # example
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [0686c2f]
+- Updated dependencies [dfc7d13]
+  - @rainbow-me/rainbowkit@0.2.3
+
+## 0.0.8
+
+### Patch Changes
+
+- 7403fab: Add dialog overlay blurs to the theme
+- Updated dependencies [4be196e]
+- Updated dependencies [7403fab]
+- Updated dependencies [4be196e]
+  - @rainbow-me/rainbowkit@0.2.2
+
 ## 0.0.7
 
 ### Patch Changes